src/requirements.txt
Django~=4.0.6
django-environ~=0.9.0
psycopg2~=2.9.3
openpyxl~=3.0.10
djangorestframework~=3.13.1
gunicorn~=20.1.0
django-filter~=22.1